# Conflicts:
#	routes/album.routes.js
#	routes/artists.routes.js
#	routes/genre.routes.js
#	routes/songs.routes.js

+++ routes/artists.routes.js
@@ -1,10 +1,40 @@
+const { isAuthenticated } = require('../middleware/middleware.js');
+// Import the models
+const Artist = require('../models/Artist.model');
+const User = require('../models/User.model');
+
 const router = require('express').Router();
 
-const Artist = require('../models/Artist.model');
+// GET all Artists
+router.get('/', async (req, res, next) => {
+    try {
+        const allArtists = await Artist.find()
+            .populate('albums')
+            .populate('genre');
+        res.status(200).json(allArtists);
+    } catch (error) {
+        next(error);
+    }
+});
 
-// create new Artist
-router.post('/', async (req, res, next) => {
+// GET one Artist
+router.get('/:artistId', async (req, res, next) => {
+    const { artistId } = req.params;
+    try {
+        const oneArtist = await Artist.findById(artistId)
+            .populate('albums')
+            .populate('genre');
+        res.status(200).json(oneArtist);
+    } catch (error) {
+        next(error);
+    }
+});
+
+// POST a new Artist
+router.post('/', isAuthenticated, async (req, res, next) => {
     const payload = req.body;
+    const { userId } = req.payload;
+    payload.createdBy = userId;
     try {
         const createdArtist = await Artist.create(payload);
         res.status(201).json(createdArtist);
@@ -14,59 +44,14 @@ router.post('/', async (req, res, next) => {
     }
 });
 
-// get all Artists
-router.get('/', async (req, res, next) => {
-    try {
-        const allArtists = await Artist.find();
-        res.status(200).json(allArtists);
-    } catch (error) {
-        next(error);
-    }
-});
-
-// get one Artist
-router.get('/:artistId', async (req, res, next) => {
-    const { artistId } = req.params;
-    try {
-        const oneArtist = await Artist.findById(artistId);
-        res.status(200).json(oneArtist);
-    } catch (error) {
-        next(error);
-    }
-});
-router.get('/', async (req, res) => {
-    try {
-        const allArtists = await Artist.find()
-            .populate('albums')
-            .populate('genre');
-
-        res.status(200).json(allArtists);
-    } catch (error) {
-        res.status(500).json({ message: 'Error getting all the Artists!!!' });
-    }
-});
-
-// get one Artist
-router.get('/:artistId', async (req, res) => {
-    const { artistId } = req.params;
-    try {
-        const oneArtist = await Artist.findById(artistId)
-            .populate('albums')
-            .populate('genre');
-
-        res.status(200).json(oneArtist);
-    } catch (error) {
-        res.status(500).json({ message: `Error getting one artist!!!` });
-    }
-});
-
-// update one Artist
-router.put('/:artistId', async (req, res, next) => {
-    const { artistId } = req.params;
+// PUT update one Artist
+router.put('/:artistId', isAuthenticated, async (req, res, next) => {
+    const { userId } = req.payload;
     const payload = req.body;
+    const { artistId } = req.params;
     try {
         const artistToUpdate = await Artist.findById(artistId);
-        if (artistToUpdate) {
+        if (artistToUpdate.createdBy == userId) {
             const updatedArtist = await Artist.findByIdAndUpdate(
                 artistId,
                 payload,
@@ -83,12 +68,18 @@ router.put('/:artistId', async (req, res, next) => {
     }
 });
 
-// delete one Artist
-router.delete('/:artistId', async (req, res, next) => {
+// DELETE Artist
+router.delete('/:artistId', isAuthenticated, async (req, res, next) => {
+    const { userId } = req.payload;
     const { artistId } = req.params;
     try {
-        const artistToDelete = await Artist.findByIdAndDelete(artistId);
-        res.status(204).json({ message: 'Artist Deleted' });
+        const artistToDelete = await Artist.findById(artistId);
+        if (artistToDelete.createdBy == userId) {
+            await Artist.findByIdAndDelete(artistId);
+            res.status(204).json();
+        } else {
+            res.status(404).json({ message: 'Genre not found' });
+        }
     } catch (error) {
         next(error);
     }
