+++ routes/songs.routes.js
@@ -1,16 +1,8 @@
+const { isAuthenticated } = require('../middleware/middleware.js');
+// Import the models
 const Songs = require('../models/Songs.model');
+const User = require('../models/User.model');
 const router = require('express').Router();
-
-// POST create new Song
-router.post('/', async (req, res, next) => {
-    const payload = req.body;
-    try {
-        const newSong = await Songs.create(payload);
-        res.status(201).json(newSong);
-    } catch (error) {
-        next(error); // Pass the error to the error handling middleware
-    }
-});
 
 // GET all Songs
 router.get('/', async (req, res, next) => {
@@ -21,7 +13,7 @@ router.get('/', async (req, res, next) => {
             .populate('genres');
         res.status(200).json(allSongs);
     } catch (error) {
-        next(error);
+        next(error); // Pass the error to the error handling middleware
     }
 });
 
@@ -39,11 +31,35 @@ router.get('/:songId', async (req, res, next) => {
     }
 });
 
-// UPDATE one Song
-router.put('/:songId', async (req, res, next) => {
+// POST create new Song
+router.post('/', isAuthenticated, async (req, res, next) => {
+    const payload = req.body;
+    const { userId } = req.payload;
+    payload.createdBy = userId;
     try {
-        const songId = req.params.songId;
-        const songToUpdate = req.body;
+        const newSong = await Songs.create(payload);
+        res.status(201).json(newSong);
+    } catch (error) {
+        console.log(error);
+        next(error);
+    }
+});
+
+// UPDATE one Song
+router.put('/:songId', isAuthenticated, async (req, res, next) => {
+    const { userId } = req.payload;
+    const songId = req.params.songId;
+    const songToUpdate = req.body;
+    try {
+        const existingSong = await Songs.findById(songId);
+        if (!existingSong) {
+            return res.status(404).json({ message: 'Song not found' });
+        }
+        if (existingSong.createdBy.toString() !== userId.toString()) {
+            return res
+                .status(403)
+                .json({ message: 'Unauthorized to update this song' });
+        }
         const updatedSong = await Songs.findByIdAndUpdate(
             songId,
             songToUpdate,
@@ -51,24 +67,24 @@ router.put('/:songId', async (req, res, next) => {
                 new: true,
             }
         );
-
-        if (!updatedSong) {
-            return res.status(404).json({ message: 'Song not found' });
-        }
+        res.status(200).json(updatedSong);
     } catch (error) {
         next(error);
     }
 });
 
-// delete one Song
-router.delete('/:songId', async (req, res, next) => {
+// DELETE one Song
+router.delete('/:songId', isAuthenticated, async (req, res, next) => {
+    const { userId } = req.payload;
+    const { songId } = req.params;
     try {
-        const songId = req.params.songId;
-        const deletedSong = await Songs.findByIdAndDelete(songId);
-        if (!deletedSong) {
-            return res.status(404).json({ message: 'Song not found.' });
+        const songToDelete = await Songs.findById(songId);
+        if (songToDelete.createdBy == userId) {
+            await Songs.findByIdAndDelete(songId);
+            res.status(204).json();
+        } else {
+            res.status(404).json({ message: 'Song not found' });
         }
-        res.status(200).send();
     } catch (error) {
         next(error);
     }
